requirements.txt
gunicorn
pandas
numpy
flask
sklearn
pyYAML
ipykernel
-e .